+++ fpsim/locations/kenya/calibrate_manual.py
@@ -20,7 +20,10 @@ do_plot_skyscrapers = True
 do_plot_cpr = True
 do_plot_tfr = True
 do_plot_pop_growth = True
-do_plot_birth_space = True
+do_plot_birth_space_afb = True
+
+# Set option to save figures
+do_save = 1
 
 
 # Set up global variables
@@ -37,13 +40,16 @@ age_bin_map = {
 
 min_age = 15
 max_age = 50
+first_birth_age = 25  # age to start assessing first birth age in model
 bin_size = 5
-mpy = 12
+mpy = 12 # months per year
 
-skyscrapers = pd.read_csv('kenya_skyscrapers.csv')
-use = pd.read_csv('use_kenya.csv')
+# Import Kenya data files to compare
+skyscrapers = pd.read_csv('kenya_skyscrapers.csv') # Age-parity distribution file
+use = pd.read_csv('use_kenya.csv') #Dichotomous contraceptive method use
+data_spaces = pd.read_csv('birth_spacing_dhs.csv')  # Birth-to-birth interval data
+data_afb = pd.read_csv('afb.table.csv')  # Ages at first birth in DHS for women age 25-50
 
-dataset = 'PMA 2022'  # Data to compare to for skyscrapers
 
 
 # Set up sim for Kenya
@@ -54,11 +60,24 @@ pars['end_year'] = 2020 # 1961 - 2020 is the normal date range
 # Free parameters for calibration
 pars['fecundity_var_low'] = 0.95
 pars['fecundity_var_high'] = 1.05
+pars['exposure_factor'] = 1
+pars['high_parity'] = 1
+pars['high_parity_nonuse'] = 1
+
+# Last free parameter, postpartum sexual activity correction or 'birth spacing preferece'
+# Set all to 1 to reset
+spacing_pars = {'space0_6': 1, 'space18_24': 1, 'space27_36': 1, 'space9_15': 1}  # output from 'optimize-space-prefs-kenya.py'
+pars['spacing_pref']['preference'][:3] = spacing_pars['space0_6']
+pars['spacing_pref']['preference'][3:6] = spacing_pars['space9_15']
+pars['spacing_pref']['preference'][6:9] = spacing_pars['space18_24']
+#pars['spacing_pref']['preference'][9:] = spacing_pars['space27_36'] # Removing this bin for Kenya as it doesn't extend out
+
+# Only other free parameters are age-based exposure and parity-based exposure, can adjust manually in kenya.py
 
 sim = fp.Sim(pars=pars)
 sim.run()
 
-
+# Plot results from sum run
 if do_plot_sim:
     sim.plot()
 
@@ -68,8 +87,15 @@ res = sim.results
 # Save people from sim
 ppl = sim.people
 
+# Set up dictionaries to compare from model vs data
+data_dict = {}
+model_dict = {} # For comparison from model to data
+
 
 def pop_growth_rate(years, population):
+        '''
+        Calculates growth rate as a time series to help compare model to data
+        '''
         growth_rate = np.zeros(len(years) - 1)
 
         for i in range(len(years)):
@@ -79,20 +105,27 @@ def pop_growth_rate(years, population):
 
         return growth_rate
 
-
+# Start series of options for plotting data to model comaprisons
 if do_plot_asfr:
+        '''
+        Plot age-specific fertility rate between model and data
+        '''
+        # Print ASFR form model in output
         for key in age_bin_map.keys():
             print(f'ASFR (annual) for age bin {key} in the last year of the sim: {res["asfr"][key][-1]}')
 
+        # Load data
         x = [1, 2, 3, 4, 5, 6, 7, 8]
         asfr_data = [3.03, 64.94, 172.12, 174.12, 136.10, 80.51, 34.88, 13.12]  # From UN Data Kenya 2020 (kenya_asfr.csv)
         x_labels = []
         asfr_model = []
 
+        # Extract from model
         for key in age_bin_map.keys():
                 x_labels.append(key)
                 asfr_model.append(res['asfr'][key][-1])
 
+        # Plot
         fig, ax = pl.subplots()
         kw = dict(lw=3, alpha=0.7, markersize=10)
         ax.plot(x, asfr_data, marker='^', color='black', label="UN data", **kw)
@@ -104,11 +137,19 @@ if do_plot_asfr:
         ax.set_ylabel('ASFR in 2019')
         ax.legend(frameon=False)
         sc.boxoff()
-        pl.savefig('figs/asfr.png')
+
+        if do_save:
+                pl.savefig('figs/asfr.png')
+
         pl.show()
 
 if do_plot_methods:
+        '''
+        Plots both dichotomous method use and non-use and contraceptive mix
+        '''
 
+        # Pull method definitions from parameters file (kenya.py)
+        # Sould be similar for other locations unless wanting to change efficacy
         methods_map_model = {  # Index, modern, efficacy
         'None': [0, False, 0.000],
         'Withdrawal': [1, False, 0.866],
@@ -123,20 +164,21 @@ if do_plot_methods:
         'Other modern': [9, True, 0.880],
         }
 
+        # Setup
         model_labels_all = list(methods_map_model.keys())
         model_labels_methods = sc.dcp(model_labels_all)
         model_labels_methods = model_labels_methods[1:]
 
         model_method_counts = sc.odict().make(keys=model_labels_all, vals=0.0)
 
-        # From model
+        # Extract from model
         for i in range(len(ppl)):
                 if ppl.alive[i] and not ppl.sex[i] and ppl.age[i] >= min_age and ppl.age[i] < max_age:
                         model_method_counts[ppl.method[i]] += 1
 
         model_method_counts[:] /= model_method_counts[:].sum()
 
-        # From data - Kenya PMA 2022 (mix_kenya.csv)
+        # Method mix from data - Kenya PMA 2022 (mix_kenya.csv)
         data_methods_mix = {
                 'Withdrawal': 1.03588605253422,
                 'Other traditional': 4.45800961894192,
@@ -149,7 +191,7 @@ if do_plot_methods:
                 'Other modern': 3.12615612282649
         }
 
-        # From data - Kenya PMA 2022 (use_kenya.csv)
+        # Method use from data - Kenya PMA 2022 (use_kenya.csv)
         data_methods_use = {
                 'No use': 51.1078954508456,
                 'Any method': 48.8921045491544
@@ -178,26 +220,32 @@ if do_plot_methods:
         df_mix = pd.DataFrame({'PMA': mix_percent_data, 'FPsim': mix_percent_model}, index=model_labels_methods)
         df_use = pd.DataFrame({'PMA': data_use_percent, 'FPsim': model_use_percent}, index=use_labels)
 
+        # Plot mix
         ax = df_mix.plot.barh(color={'PMA':'black', 'FPsim':'cornflowerblue'})
         ax.set_xlabel('Percent users')
         ax.set_title('Contraceptive method mix model vs data')
+        if do_save:
+                pl.savefig("figs/method_mix.png", bbox_inches='tight', dpi=100)
 
-        pl.savefig("figs/method_mix.png", bbox_inches='tight', dpi=100)
-
+        # Plot use
         ax = df_use.plot.barh(color={'PMA':'black', 'FPsim':'cornflowerblue'})
         ax.set_xlabel('Percent')
         ax.set_title('Contraceptive method use model vs data')
-
-        pl.savefig("figs/method_use.png", bbox_inches='tight', dpi=100)
-
-        #pl.show()
+        if do_save:
+                pl.savefig("figs/method_use.png", bbox_inches='tight', dpi=100)
 
 
 if do_plot_skyscrapers:
+        '''
+        Plot an age-parity distribution for model vs data
+        '''
 
+        dataset = 'PMA 2022'  # Data to compare to for skyscrapers, can also use DHS 2014
+
+        # Set up
         age_keys = list(age_bin_map.keys())[1:]
         age_bins = pl.arange(min_age, max_age, bin_size)
-        parity_bins = pl.arange(0, 7)
+        parity_bins = pl.arange(0, 7) # Plot up to parity 6
         n_age = len(age_bins)
         n_parity = len(parity_bins)
         x_age = pl.arange(n_age)
@@ -237,6 +285,7 @@ if do_plot_skyscrapers:
         for key in ['Data', 'Model']:
                 sky_arr[key] /= sky_arr[key].sum() / 100
 
+        # Find diff to help visualize in plotting
         sky_arr['Diff: data - model'] = sky_arr['Data']-sky_arr['Model']
 
         # Plot skyscrapers
@@ -254,16 +303,23 @@ if do_plot_skyscrapers:
                 pl.gca().view_init(30, 45)
                 pl.draw()
 
-                pl.savefig('figs/skyscrapers_' + str(key.lower()) + '.png')
+
+                if do_save:
+                        pl.savefig('figs/skyscrapers_' + str(key.lower()) + '.png')
 
                 pl.show()
 
+
 if do_plot_cpr:
+        '''
+        Plot contraceptive prevalence rate for model vs data
+        '''
 
         # Import data
         data_cpr = pd.read_csv('kenya_cpr.csv') # From UN Data Portal
         data_cpr = data_cpr[data_cpr['year'] <= 2020] # Restrict years to plot
 
+        # Plot
         pl.plot(data_cpr['year'], data_cpr['cpr'], label='UN Data Portal', color='black')
         pl.plot(res['t'], res['cpr']*100, label='FPsim', color='cornflowerblue')
         pl.xlabel('Year')
@@ -271,14 +327,21 @@ if do_plot_cpr:
         pl.title('Contraceptive Prevalence Rate in Data vs Model - Kenya')
         pl.legend()
 
-        pl.savefig('figs/cpr_over_sim.png')
+        if do_save:
+                pl.savefig('figs/cpr_over_sim.png')
+
         pl.show()
 
+
 if do_plot_tfr:
+        '''
+        Plot total fertility rate for model vs data
+        '''
 
         # Import data
         data_tfr = pd.read_csv('kenya_tfr.csv')
 
+        # Plot
         pl.plot(data_tfr['year'], data_tfr['tfr'], label='World Bank', color='black')
         pl.plot(res['tfr_years'], res['tfr_rates'], label='FPsim', color='cornflowerblue')
         pl.xlabel('Year')
@@ -286,20 +349,28 @@ if do_plot_tfr:
         pl.title('Total Fertility Rate in Data vs Model - Kenya')
         pl.legend()
 
-        pl.savefig('figs/tfr_over_sim.png')
+        if do_save:
+                pl.savefig('figs/tfr_over_sim.png')
+
         pl.show()
 
 if do_plot_pop_growth:
+        '''
+        Plot annual population growth rate for model vs data
+        '''
 
+        # Import data
         data_popsize = pd.read_csv('kenya_popsize.csv')
         data_popsize = data_popsize[data_popsize['year'] <= 2020]  # Restrict years to plot
 
         data_pop_years = data_popsize['year'].to_numpy()
         data_population = data_popsize['population'].to_numpy()
 
+        # Extract from model
         model_growth_rate = pop_growth_rate(res['tfr_years'], res['pop_size'])
         data_growth_rate = pop_growth_rate(data_pop_years, data_population)
 
+        # Plot
         pl.plot(data_pop_years[1:], data_growth_rate, label='World Bank', color='black')
         pl.plot(res['tfr_years'][1:], model_growth_rate, label='FPsim', color='cornflowerblue')
         pl.xlabel('Year')
@@ -307,61 +378,91 @@ if do_plot_pop_growth:
         pl.title('Population Growth Rate Data vs Model - Kenya')
         pl.legend()
 
-        pl.savefig('figs/popgrowth_over_sim.png')
+        if do_save:
+                pl.savefig('figs/popgrowth_over_sim.png')
+
         pl.show()
 
-if do_plot_birth_space:
 
+if do_plot_birth_space_afb:
+        '''
+        Plot birth space and age at first birth for model vs data
+        '''
+
+        # Set up
         spacing_bins = sc.odict({'0-12': 0, '12-24': 1, '24-48': 2, '>48': 4})  # Spacing bins in years
-
         model_age_first = []
         model_spacing = []
         model_spacing_counts = sc.odict().make(keys=spacing_bins.keys(), vals=0.0)
+        data_spacing_counts = sc.odict().make(keys=spacing_bins.keys(), vals=0.0)
+
+
+        # Extract age at first birth and birth spaces from model
         for i in range(len(ppl)):
                 if ppl.alive[i] and not ppl.sex[i] and ppl.age[i] >= min_age and ppl.age[i] < max_age:
-                        if len(ppl.dobs[i]):
+                        if len(ppl.dobs[i]) == 0:
+                                model_age_first.append(float('inf'))
+                        if len(ppl.dobs[i]) and ppl.age[i] >= first_birth_age:
                                 model_age_first.append(ppl.dobs[i][0])
                         if len(ppl.dobs[i]) > 1:
                                 for d in range(len(ppl.dobs[i]) - 1):
                                         space = ppl.dobs[i][d + 1] - ppl.dobs[i][d]
                                         ind = sc.findinds(space > spacing_bins[:])[-1]
                                         model_spacing_counts[ind] += 1
-
                                         model_spacing.append(space)
 
-        spacing = pd.DataFrame(data=model_spacing)
-        ax = spacing.plot.kde()
-        ax.set_xlabel('Years')
-        ax.set_ylabel('Density')
-        ax.set_title('Birth space in Model')
-        pl.show()
-
-        age_first = pd.DataFrame(data=model_age_first)
-        ax = age_first.plot.kde()
-        ax.set_xlabel('Years')
-        ax.set_ylabel('Density')
-        ax.set_title('Age at first birth in Model')
-        pl.show()
-
-
-
+        # Normalize model birth space bin counts to percentages
         model_spacing_counts[:] /= model_spacing_counts[:].sum()
         model_spacing_counts[:] *= 100
-        print(f'Model spacing counts: {model_spacing_counts}')
 
-        model_spacing_stats = np.array([np.percentile(model_spacing, 25),
-                                                np.percentile(model_spacing, 50),
-                                                np.percentile(model_spacing, 75)])
-        model_age_first_stats = np.array([np.percentile(model_age_first, 25),
-                                                  np.percentile(model_age_first, 50),
-                                                  np.percentile(model_age_first, 75)])
+        age_first_birth_model = pd.DataFrame(data=model_age_first)
 
-        space_df = pd.DataFrame(data=model_spacing_counts.values(), index=model_spacing_counts.keys())
-        ax = space_df.plot.barh(color='cornflowerblue')
-        ax.set_title('Birth spacing bins in Model vs Data')
-        ax.set_ylabel('Space between births in months')
-        ax.set_xlabel('Percent of all births')
-        pl.savefig('figs/birth_space_bins.png')
+        # Extract birth spaces and age at first birth from data
+        for i, j in data_spaces.iterrows():
+                space = j['space_mo'] / mpy
+                ind = sc.findinds(space > spacing_bins[:])[-1]
+                data_spacing_counts[ind] += j['Freq']
+
+        age_first_birth_data = pd.DataFrame(data=data_afb)
+
+        # Normalize dat birth space bin counts to percentages
+        data_spacing_counts[:] /= data_spacing_counts[:].sum()
+        data_spacing_counts[:] *= 100
+
+        # Plot age at first birth (histogram with KDE)
+        sns.histplot(data=age_first_birth_model, stat='proportion', kde=True, binwidth=1, color='cornflowerblue', label='FPsim')
+        sns.histplot(x=age_first_birth_data['afb'], stat='proportion', kde=True, weights=age_first_birth_data['wt'], binwidth=1, color='dimgrey', label='DHS data')
+        pl.xlabel('Age at first birth')
+        pl.legend()
+
+        if do_save:
+                pl.savefig('figs/age_first_birth.png', bbox_inches='tight', dpi=100)
+
+        pl.show()
+
+
+        # Plot birth space bins with diff
+        data_dict['spacing_bins'] = np.array(data_spacing_counts.values())
+        model_dict['spacing_bins'] = np.array(model_spacing_counts.values())
+
+        diff = model_dict['spacing_bins'] - data_dict['spacing_bins']
+
+        res_bins = np.array([[model_dict['spacing_bins']], [data_dict['spacing_bins']], [diff]])
+
+        bins_frame = pd.DataFrame(
+                {'Model': model_dict['spacing_bins'], 'Data': data_dict['spacing_bins'], 'Diff': diff},
+                index=spacing_bins.keys())
+
+        print(bins_frame) # Print in output, remove if not needed
+
+        ax = bins_frame.plot.barh(color={'Data': 'black', 'Model': 'cornflowerblue', 'Diff': 'red'})
+        ax.set_xlabel('Percent of live birth spaces')
+        ax.set_ylabel('Birth space in months')
+        ax.set_title('Birth space bins calibration - Kenya')
+
+        if do_save:
+                pl.savefig('figs/birth_space_bins_kenya.png', bbox_inches='tight', dpi=100)
+
         pl.show()
 
 sc.toc()